server/requirements.txt
python-dotenv
flask
flask-bcrypt
flask-cors
flask-session
redis
pony
